+++ tests/phpunit/lib/DDBTestConfig.class.inc
@@ -20,6 +20,8 @@ class DDBTestConfig {
     $this->user = isset($_SERVER['DDBTEST_USER']) ? $_SERVER['DDBTEST_USER'] : '3206596642';
     $this->pass = isset($_SERVER['DDBTEST_PASS']) ? $_SERVER['DDBTEST_PASS'] : '12345';
     $this->lms = isset($_SERVER['DDBTEST_LMS']) ? $_SERVER['DDBTEST_LMS'] : '';
+    $this->screenshotPath = isset($_SERVER['DDBTEST_SCREENSHOT_PATH']) ? $_SERVER['DDBTEST_SCREENSHOT_PATH'] : '';
+    $this->screenshotUrl = isset($_SERVER['DDBTEST_SCREENSHOT_URL']) ? $_SERVER['DDBTEST_SCREENSHOT_URL'] : '';
   }
 
   /**
@@ -74,5 +76,13 @@ class DDBTestConfig {
 
   public function getLms() {
     return $this->lms;
+  }
+
+  public function getScreenshotPath() {
+    return $this->screenshotPath;
+  }
+
+  public function getScreenshotUrl() {
+    return $this->screenshotUrl;
   }
 }
