+++ tests/phpunit/tests/CollectionViewTest.php
@@ -1,17 +1,10 @@
 <?php
 
-require_once(__DIR__ . '/../bootstrap.php');
+require_once 'Ding2TestBase.php';
 
-class CollectionViewTest extends PHPUnit_Extensions_SeleniumTestCase {
-  protected $abstractedPage;
-  protected $config;
-
+class CollectionViewTest extends Ding2TestBase {
   protected function setUp() {
-    $this->abstractedPage = new DDBTestPageAbstraction($this);
-    $this->config = new DDBTestConfig();
-
-    $this->setBrowser($this->config->getBrowser());
-    $this->setBrowserUrl($this->config->getUrl());
+    parent::setUp();
     resetState($this->config->getLms());
   }
 
@@ -75,8 +68,8 @@ class CollectionViewTest extends PHPUnit_Extensions_SeleniumTestCase {
     }
     else {
       $this->assertTrue($this->isElementPresent('css=div.ding-popup-content .messages.status'));
-     } 
-    
+     }
+
     // Test if this is a collection page.
     // If this is truly a collection item is should have certain
     // information about item types and the quantity of those.
@@ -87,7 +80,7 @@ class CollectionViewTest extends PHPUnit_Extensions_SeleniumTestCase {
     $this->click('link=Bog (2)');
     sleep(5);
     $this->assertTrue((bool) preg_match('/^.*ting\/collection\/870970-basis%3A51043138#Bog$/', $this->getLocation()));
-  //   $this->click('link=Ebog (1)');          
+  //   $this->click('link=Ebog (1)');
   //   $this->assertTrue((bool) preg_match('/^.*ting\/collection\/870970-basis%3A51043138#Ebog$/', $this->getLocation()));
   }
 
@@ -102,5 +95,5 @@ class CollectionViewTest extends PHPUnit_Extensions_SeleniumTestCase {
     $this->abstractedPage->userLogin($this->config->getUser(), $this->config->getPass());
     $this->testCollectionViewAnonymous();
   }
-  
+
 }
