+++ tests/phpunit/tests/ItemPageRelationsTest.php
@@ -1,17 +1,8 @@
 <?php
 
-class ItemPageRelationsTest extends PHPUnit_Extensions_SeleniumTestCase {
-  protected $abstractedPage;
-  protected $config;
+require_once 'Ding2TestBase.php';
 
-  protected function setUp() {
-    $this->abstractedPage = new DDBTestPageAbstraction($this);
-    $this->config = new DDBTestConfig();
-
-    $this->setBrowser($this->config->getBrowser());
-    $this->setBrowserUrl($this->config->getUrl());
-  }
-
+class ItemPageRelationsTest extends Ding2TestBase {
   /**
    * Test related materials of a certain item as anonymous.
    *
